Changed int to floating point num
- replaced int with num, now it's all floating point and possibly terrible
- updated readme and examples

+++ interpretation/functionStack.hpp
@@ -8,7 +8,7 @@
 #include "../content/function.hpp"
 #include "../content/variables.hpp"
 
-Variable* NewVariable(std::string& token, long long value);
+Variable* NewVariable(std::string& token, long double value);
 Variable* NewVariable(std::string& token, std::string& value);
 
 struct InstanceLevel {
